avoid circular refrences between base.py and stellantis.py
