+++ custom_components/stellantis_vehicles/base.py
@@ -1,12 +1,8 @@
 import logging
 import re
-from datetime import datetime, timedelta, UTC
-import json
+from datetime import datetime
 
-from homeassistant.core import HomeAssistant
-from homeassistant.config_entries import ConfigEntry
-
-from homeassistant.helpers.update_coordinator import ( CoordinatorEntity, DataUpdateCoordinator )
+from homeassistant.helpers.update_coordinator import CoordinatorEntity
 from homeassistant.components.device_tracker import ( SourceType, TrackerEntity )
 from homeassistant.components.sensor import RestoreSensor, SensorEntityDescription
 from homeassistant.components.binary_sensor import BinarySensorEntity
@@ -17,289 +13,18 @@ from homeassistant.components.text import TextEntity
 from homeassistant.core import callback
 from homeassistant.helpers.restore_state import RestoreEntity
 from homeassistant.const import ( STATE_UNAVAILABLE, STATE_UNKNOWN, STATE_ON, STATE_OFF )
-from homeassistant.exceptions import ConfigEntryAuthFailed
 
-from .stellantis import StellantisVehicles
-
-from .utils import ( date_from_pt_string, get_datetime, timestring_to_datetime )
+from .utils import ( get_datetime, timestring_to_datetime )
 
 from .const import (
     DOMAIN,
     FIELD_MOBILE_APP,
-    VEHICLE_TYPE_ELECTRIC,
     VEHICLE_TYPE_HYBRID,
-    UPDATE_INTERVAL
 )
 
+from .stellantis import StellantisVehicleCoordinator
+
 _LOGGER = logging.getLogger(__name__)
-
-class StellantisVehicleCoordinator(DataUpdateCoordinator):
-    def __init__(
-            self, 
-            hass: HomeAssistant, 
-            config: ConfigEntry, 
-            vehicle, 
-            stellantis: StellantisVehicles, 
-            translations
-        ):
-        super().__init__(hass, _LOGGER, name = DOMAIN, update_interval=timedelta(seconds=UPDATE_INTERVAL))
-
-        self._hass = hass
-        self._translations = translations
-        self._config = config
-        self._vehicle = vehicle
-        self._stellantis = stellantis
-        self._data = {}
-        self._sensors = {}
-        self._commands_history = {}
-        self._disabled_commands = []
-        self._last_trip = None
-#        self._total_trip = None
-
-    async def _async_update_data(self):
-        _LOGGER.debug("---------- START _async_update_data")
-        try:
-            # Vehicle status
-            self._data = await self._stellantis.get_vehicle_status()
-        except ConfigEntryAuthFailed as e:
-            raise ConfigEntryAuthFailed from e
-        except Exception as e:
-            _LOGGER.error(str(e))
-        _LOGGER.debug(self._config)
-        _LOGGER.debug(self._data)
-        await self.after_async_update_data()
-        _LOGGER.debug("---------- END _async_update_data")
-
-    def get_translation(self, path, default = None):
-        return self._translations.get(path, default)
-
-    @property
-    def vehicle_type(self):
-        return self._stellantis.get_config("type")
-
-    @property
-    def command_history(self):
-        history = {}
-        if not self._commands_history:
-            return history
-        reorder_actions = list(reversed(self._commands_history.keys()))
-        for action_id in reorder_actions:
-            action_name = self._commands_history[action_id]["name"]
-            action_updates = self._commands_history[action_id]["updates"]
-            reorder_updates = reversed(action_updates)
-            for update in reorder_updates:
-                status = update["info"]
-                translation_path = f"component.stellantis_vehicles.entity.sensor.command_status.state.{status}"
-                status = self.get_translation(translation_path, status)
-                history.update({update["date"].strftime("%d/%m/%y %H:%M:%S:%f")[:-4]: str(action_name) + ": " + status})
-        return history
-
-    @property
-    def pending_action(self):
-        if not self._commands_history:
-            return False
-        last_action_id = list(self._commands_history.keys())[-1]
-        return not self._commands_history[last_action_id]["updates"]
-
-    async def update_command_history(self, action_id, update = None):
-        if action_id not in self._commands_history:
-            return
-        if update:
-            self._commands_history[action_id]["updates"].append({"info": update, "date": get_datetime()})
-            if update == "99":
-                self._disabled_commands.append(self._commands_history[action_id]["name"])
-        self.async_update_listeners()
-
-    async def send_command(self, name, service, message):
-        action_id = await self._stellantis.send_mqtt_message(service, message)
-        self._commands_history.update({action_id: {"name": name, "updates": []}})
-
-    async def send_wakeup_command(self, button_name):
-        await self.send_command(button_name, "/VehCharge/state", {"action": "state"})
-
-    async def send_doors_command(self, button_name):
-        current_status = self._sensors["doors"]
-        new_status = "lock"
-        if current_status == "Locked":
-            new_status = "unlock"
-        await self.send_command(button_name, "/Doors", {"action": new_status})
-
-    async def send_horn_command(self, button_name):
-        await self.send_command(button_name, "/Horn", {"nb_horn": "2", "action": "activate"})
-
-    async def send_lights_command(self, button_name):
-        await self.send_command(button_name, "/Lights", {"duration": "10", "action": "activate"})
-
-    async def send_charge_command(self, button_name):
-        current_hour = self._sensors["battery_charging_time"]
-        current_status = self._sensors["battery_charging"]
-        charge_type = "immediate"
-        if current_status == "InProgress":
-            charge_type = "delayed"
-        await self.send_command(button_name, "/VehCharge", {"program": {"hour": current_hour.hour, "minute": current_hour.minute}, "type": charge_type})
-
-    def get_programs(self):
-        default_programs = {
-           "program1": {"day": [0, 0, 0, 0, 0, 0, 0], "hour": 34, "minute": 7, "on": 0},
-           "program2": {"day": [0, 0, 0, 0, 0, 0, 0], "hour": 34, "minute": 7, "on": 0},
-           "program3": {"day": [0, 0, 0, 0, 0, 0, 0], "hour": 34, "minute": 7, "on": 0},
-           "program4": {"day": [0, 0, 0, 0, 0, 0, 0], "hour": 34, "minute": 7, "on": 0}
-        }
-#        active_programs = None
-        if "programs" in self._data["preconditionning"]["airConditioning"]:
-            current_programs = self._data["preconditionning"]["airConditioning"]["programs"]
-            if current_programs:
-                for program in current_programs:
-                    if program:
-                        if program and program.get("occurence", {}).get("day", None) and program.get("start", None):
-                            date = date_from_pt_string(program["start"])
-                            config = {
-                                "day": [
-                                    int("Mon" in program["occurence"]["day"]),
-                                    int("Tue" in program["occurence"]["day"]),
-                                    int("Wed" in program["occurence"]["day"]),
-                                    int("Thu" in program["occurence"]["day"]),
-                                    int("Fri" in program["occurence"]["day"]),
-                                    int("Sat" in program["occurence"]["day"]),
-                                    int("Sun" in program["occurence"]["day"])
-                                ],
-                                "hour": date.hour,
-                                "minute": date.minute,
-                                "on": int(program["enabled"])
-                            }
-                            default_programs["program"+str(program["slot"])] = config
-        return default_programs
-
-    async def send_air_conditioning_command(self, button_name):
-        current_status = self._sensors["air_conditioning"]
-        new_status = "activate"
-        if current_status == "Enabled":
-            new_status = "deactivate"
-        await self.send_command(button_name, "/ThermalPrecond", {"asap": new_status, "programs": self.get_programs()})
-
-    async def send_abrp_data(self):
-        tlm = {
-            "utc": int(get_datetime().astimezone(UTC).timestamp()),
-            "soc": None,
-            "power": None,
-            "speed": None,
-            "lat": None,
-            "lon": None,
-            "is_charging": False,
-            "is_dcfc": False,
-            "is_parked": False
-        }
-
-        if "battery" in self._sensors:
-            tlm["soc"] = self._sensors["battery"]
-        if "kinetic" in self._data and "speed" in self._data["kinetic"]:
-            tlm["speed"] = self._data["kinetic"]["speed"]
-        if "lastPosition" in self._data:
-            tlm["lat"] = float(self._data["lastPosition"]["geometry"]["coordinates"][1])
-            tlm["lon"] = float(self._data["lastPosition"]["geometry"]["coordinates"][0])
-        if "battery_charging" in self._sensors:
-            tlm["is_charging"] = self._sensors["battery_charging"] == "InProgress"
-        if "battery_charging_type" in self._sensors:
-            tlm["is_dcfc"] = tlm["is_charging"] and self._sensors["battery_charging_type"] == "Quick"
-        if "battery_soh" in self._sensors and self._sensors["battery_soh"]:
-            tlm["soh"] = float(self._sensors["battery_soh"])
-        if "lastPosition" in self._data and "properties" in self._data["lastPosition"] and "heading" in self._data["lastPosition"]["properties"]:
-            tlm["heading"] = float(self._data["lastPosition"]["properties"]["heading"])
-        if "lastPosition" in self._data and len(self._data["lastPosition"]["geometry"]["coordinates"]) == 3:
-            tlm["elevation"] = float(self._data["lastPosition"]["geometry"]["coordinates"][2])
-        if "temperature" in self._sensors:
-            tlm["ext_temp"] = self._sensors["temperature"]
-        if "mileage" in self._sensors:
-            tlm["odometer"] = self._sensors["mileage"]
-        if "autonomy" in self._sensors:
-            tlm["est_battery_range"] = self._sensors["autonomy"]
-
-        params = {"tlm": json.dumps(tlm), "token": self._sensors["text_abrp_token"]}
-        await self._stellantis.send_abrp_data(params)
-
-
-    async def after_async_update_data(self):
-        if self.vehicle_type in [VEHICLE_TYPE_ELECTRIC, VEHICLE_TYPE_HYBRID]:
-            if not hasattr(self, "_manage_charge_limit_sent"):
-                self._manage_charge_limit_sent = False
-
-            if "battery_charging" in self._sensors:
-                if self._sensors["battery_charging"] == "InProgress" and not self._manage_charge_limit_sent:
-                    charge_limit_on = "switch_battery_charging_limit" in self._sensors and self._sensors["switch_battery_charging_limit"]
-                    charge_limit = None
-                    if "number_battery_charging_limit" in self._sensors and self._sensors["number_battery_charging_limit"]:
-                        charge_limit = self._sensors["number_battery_charging_limit"]
-                    if charge_limit_on and charge_limit and "battery" in self._sensors:
-                        current_battery = self._sensors["battery"]
-                        if int(float(current_battery)) >= int(charge_limit):
-                            button_name = self.get_translation("component.stellantis_vehicles.entity.button.charge_start_stop.name")
-                            await self.send_charge_command(button_name)
-                            self._manage_charge_limit_sent = True
-                elif self._sensors["battery_charging"] != "InProgress" and not self._manage_charge_limit_sent:
-                    self._manage_charge_limit_sent = False
-
-            if "switch_abrp_sync" in self._sensors and self._sensors["switch_abrp_sync"] and "text_abrp_token" in self._sensors and len(self._sensors["text_abrp_token"]) == 36:
-                await self.send_abrp_data()
-
-        if "engine" in self._sensors and "ignition" in self._data and "type" in self._data["ignition"]:
-            current_engine_status = self._sensors["engine"]
-            new_engine_status = self._data["ignition"]["type"]
-            if current_engine_status != "Stop" and new_engine_status == "Stop":
-                await self.get_vehicle_last_trip()
-
-        if "number_refresh_interval" in self._sensors and self._sensors["number_refresh_interval"] > 0 and self._sensors["number_refresh_interval"] != self._update_interval_seconds:
-            self.update_interval = timedelta(seconds=self._sensors["number_refresh_interval"])
-            self._stellantis._refresh_interval = self._sensors["number_refresh_interval"]
-
-    async def get_vehicle_last_trip(self):
-        trips = await self._stellantis.get_vehicle_last_trip()
-        if "_embedded" in trips and "trips" in trips["_embedded"] and trips["_embedded"]["trips"]:
-            if not self._last_trip or self._last_trip["id"] != trips["_embedded"]["trips"][-1]["id"]:
-                self._last_trip = trips["_embedded"]["trips"][-1]
-
-#     def parse_trips_page_data(self, data):
-#         result = []
-#         if "_embedded" in data and "trips" in data["_embedded"] and data["_embedded"]["trips"]:
-#             for trip in data["_embedded"]["trips"]:
-#                 item = {"engine": {}}
-#                 if "startMileage" in trip:
-#                     item["start_mileage"] = float(trip["startMileage"])
-#                 if "energyConsumptions" in trip:
-#                     for consuption in trip["energyConsumptions"]:
-#                         if not "type" in consuption or not "consumption" in consuption or not "avgConsumption" in consuption:
-#                             _LOGGER.error(consuption)
-#                             continue
-#                         trip_consumption = float(consuption["consumption"]) / 1000
-#                         trip_avg_consumption = float(consuption["avgConsumption"]) / 1000
-#                         if trip_consumption <= 0 or trip_avg_consumption <= 0:
-#                             _LOGGER.error(consuption)
-#                             continue
-#                         trip_distance = trip_consumption / (trip_avg_consumption / 100)
-#                         item["engine"][consuption["type"].lower()] = {
-#                             "distance": trip_distance,
-#                             "consumption": trip_consumption
-#                         }
-#                 else:
-#                     continue
-#                 if not item["engine"]:
-#                     continue
-#                 result.append(item)
-#         return result
-#
-#     async def get_vehicle_trips(self):
-#         vehicle_trips_request = await self._stellantis.get_vehicle_trips()
-#         total_trips = int(vehicle_trips_request["total"])
-#         next_page_url = vehicle_trips_request["_links"]["next"]["href"]
-#         pages = math.ceil(total_trips / 60) - 1
-#         result = self.parse_trips_page_data(vehicle_trips_request)
-#         if pages > 1:
-#             for _ in range(pages):
-#                 page_token = next_page_url.split("pageToken=")[1]
-#                 page_trips_request = await self._stellantis.get_vehicle_trips(page_token)
-#                 result = result + self.parse_trips_page_data(page_trips_request)
-#                 if "next" in page_trips_request["_links"]:
-#                     next_page_url = page_trips_request["_links"]["next"]["href"]
-#         self._total_trip = {"totals": total_trips, "trips": result}
 
 class StellantisBaseEntity(CoordinatorEntity):
     def __init__(
